requirements.txt
torch torchvision
transformers
datasets
scikit-learn
matplotlib seaborn
pillow
tqdm